Preparations for reworking the memstick folder move functionality
