+++ Core/Core.vcxproj.filters
@@ -1303,6 +1303,9 @@
     <ClCompile Include="Util\GameDB.cpp">
       <Filter>Util</Filter>
     </ClCompile>
+    <ClCompile Include="Util\MemStick.cpp">
+      <Filter>Util</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="ELF\ElfReader.h">
@@ -2080,6 +2083,9 @@
       <Filter>MIPS\ARM64</Filter>
     </ClInclude>
     <ClInclude Include="Util\GameDB.h">
+      <Filter>Util</Filter>
+    </ClInclude>
+    <ClInclude Include="Util\MemStick.h">
       <Filter>Util</Filter>
     </ClInclude>
   </ItemGroup>
